Add Dockerfile and make docker- target for containerized builds
This change adds a basic Dockerfile that includes the c-for-go
dependency to allow the generation of bindings without installing
this dependency.

A generic make target `docker-%` is added to run any of the recognised
make targets in the container.

Note that this pins:
* The GO version to 1.14.4
* The c-for-go commit to 8eeee8c3b71f9c3c90c4a73db54ed08b0bba971d

To ensure that this does not introduce any changes when generating
the bindings.

Signed-off-by: Evan Lezar <elezar@nvidia.com>