Use sed instead of patch to update header file
This change removes the need to modify the nvml.h.patch file when
updating nvml.h. The target header file is generated by running a sed
command over the input file to reformat the typedef structs.

This also applies the changes to the generated files in pkg making some
whitespace changes in `pkg/nvml.h` and updating generation timestamps in
go file.

Signed-off-by: Evan Lezar <elezar@nvidia.com>