Remove funcs attached to return type
Having these here makes it hard to mock this type, and turning it into
an interface (at this level) is hard because then you can't properly
compare it to the const definition for SUCCESS, NOT_SUPPORTED, etc.
(without a cast).

Until we come up with a better way to support mocking this type with
funcs attached to it, it's better to just have the top level function
defined (which will never change / go away.

Signed-off-by: Kevin Klues <kklues@nvidia.com>