Fix GetComputeRunningProcesses on CUDA 10.x
Due to additional fields added to the nvmlProcessInfo_st in CUDA 11.x there
is a mismatch in the size of the elements of the slice returned by the
NVML call to get the running Compute / Graphics processes. This change
attempts to detect whether the returned slice needs to be reinterpreted
as a slice of smaller elements.

Signed-off-by: Evan Lezar <elezar@nvidia.com>