Add patch to pull nested anonymous structs out of structs and unions
Newer versionf of NVML have introduced anonymous structs inside of versious
top-level structs and unions, and godefs doesn't handle these unfortunately. To
compensate for this, we use a coccinelle file to restructure nvml.h and create
well-defined structs (with typedefs) instead of anonymous structs. These new
typedefs are named such that they can be easily referred to in user-code as
desired (e.g. to populate them and then cast the original union type to them).

Signed-off-by: Kevin Klues <kklues@nvidia.com>