Rename type used to provide indirection for `V()` style methods
This is technically a breaking change, but I can't imagine there being anyone
who creates a variable to one of these types. If they do, I imagine they use
the `:=` syntax and not the explicit `var` syntax, so they won't be naming the
tspe anyway. I'm OK breaking the 1 person who this might affect.

Signed-off-by: Kevin Klues <kklues@nvidia.com>